requirements.txt
python-telegram-bot[job-queue]
asyncpg
aiohttp
pyairtable
python-dotenv
